chore(grammargen): gate diagnostic probes
