+++ grammargen/z_diag_compare_test.go
@@ -1,3 +1,5 @@
+//go:build diagnostic
+
 package grammargen
 
 import (
