+++ grammargen/z_diag_generate_fastpath_test.go
@@ -1,3 +1,5 @@
+//go:build diagnostic
+
 package grammargen
 
 import (
