+++ grammargen/z_diag_ini_test.go
@@ -1,3 +1,5 @@
+//go:build diagnostic
+
 package grammargen
 
 import (
