+++ grammargen/z_diag_scala_root_test.go
@@ -1,3 +1,5 @@
+//go:build diagnostic
+
 package grammargen
 
 import (
