+++ grammargen/z_diag_yaml_root_eof_test.go
@@ -1,3 +1,5 @@
+//go:build diagnostic
+
 package grammargen
 
 import (
